+++ common/constants/constCodes.js
@@ -1,9 +1,12 @@
-const errorCodes = {
+const constCodes = {
+    'ok': 200,
     'created': 201,
     'deleted': 202,
     'badRequest': 400,
+    'unauthorized': 401,
+    'forbidden': 403,
     'notFound': 404,
     'duplicatedField': 11000,
 };
 
-export default errorCodes;
+export default constCodes;