+++ common/constants/constMessages.js
@@ -1,0 +1,18 @@
+const constMessages = {
+    'accessorieNotExists': 'Accessorie not exists',
+    'admin': 'admin',
+    'adminAdded': 'Admin has been added!',
+    'appIsRunning': 'App is running on port',
+    'userNotExists': 'User not exists',
+    'assetNotExists': 'Asset not exists',
+    'actionSucceed': 'OK',
+    'generatedUsers': 'Generated users',
+    'duplicatedField': 'Duplicated field',
+    'internalServerError': 'Internal server error',
+    'invalidPermissions': 'Invalid permissions',
+    'invalidCredentials': 'Invalid credentials',
+    'usernameAndPasswordAreRequired': 'Username and password are required',
+    'tokenMissing': 'Token missing'
+};
+
+export default constMessages;