+++ common/constants/errorMessages.js
@@ -1,8 +1,0 @@
-const errorMessages = {
-    'accessorieNotExists': 'Accessorie not exists',
-    'actionSucceed': 'OK',
-    'duplicatedField': 'Duplicated field',
-    'internalServerError': 'Internal server error'
-};
-
-export default errorMessages;