+++ common/schemas/assetsSchema.js
@@ -1,0 +1,30 @@
+import mongoose from 'mongoose';
+
+const schema = new mongoose.Schema({
+    id: {
+        type: Number,
+        unique: [true, 'Asset already exists'],
+        immutable: true,
+    },
+    name: {
+        type: String,
+        required: [true, 'Name is required'],
+        unique: [true, 'Name already exists'],
+        minlength: [2, 'Name is shorter than the minimum allowed length (2)'],
+        maxlength: [31, 'Name is longer than the maximum allowed length (31)'],
+    },
+    isDeleted: {
+        type: Boolean,
+        default: false,
+    },
+    assigneeModel: {
+        type: String,
+        enum: ['users', 'locations'],
+    },
+    assignee: {
+        type: mongoose.Schema.Types.Number,
+        ref: 'assigneeModel',
+    },
+}, { versionKey: false });
+
+export default schema;