+++ common/schemas/usersSchema.js
@@ -1,0 +1,88 @@
+import mongoose from 'mongoose';
+import bcrypt from 'bcrypt';
+
+const schema = new mongoose.Schema({
+    id: {
+        type: Number,
+        unique: [true, 'User already exists'],
+        immutable: true,
+    },
+    username: {
+        type: String,
+        required: [true, 'Username is required'],
+        minlength: [2, 'Username is shorter than the minimum allowed length (2)'],
+        maxlength: [31, 'Username is longer than the maximum allowed length (31)'],
+    },
+    password: {
+        type: String,
+        required: [true, 'Password is required'],
+        minlength: [8, 'Password is shorter than the minimum allowed length (2)'],
+        maxlength: [64, 'Password is longer than the maximum allowed length (64)'],
+    },
+    email: {
+        type: String,
+        required: true,
+        match: [/^\S+@\S+\.\S+$/, 'Invalid form'],
+        maxlength: [127, 'Email is longer than the maximum allowed length (31)'],
+    },
+    firstName: {
+        type: String,
+        required: [true, 'Firstname is required'],
+        minlength: [2, 'Firstname is shorter than the minimum allowed length (2)'],
+        maxlength: [31, 'Firstname is longer than the maximum allowed length (31)'],
+    },
+    lastName: {
+        type: String,
+        required: [true, 'Lastname is required'],
+        minlength: [2, 'Lastname is shorter than the minimum allowed length (2)'],
+        maxlength: [31, 'Lastname is longer than the maximum allowed length (31)'],
+    },
+    middleName: {
+        type: String,
+        minlength: [2, 'Middlename is shorter than the minimum allowed length (2)'],
+        maxlength: [31, 'Middlename is longer than the maximum allowed length (31)'],
+    },
+    role: {
+        type: String,
+        enum: {
+            values: ['user', 'admin'],
+            message: 'Role must be user or admin',
+        },
+        default: 'user',
+    },
+    location: {
+        type: mongoose.Schema.Types.Number,
+        ref: 'locations',
+    },
+    company: {
+        type: mongoose.Schema.Types.Number,
+        ref: 'departments',
+    },
+    department: {
+        type: mongoose.Schema.Types.Number,
+        ref: 'departments',
+    },
+    isRemote: {
+        type: Boolean,
+    },
+    isRemembered: {
+        type: Boolean,
+    },
+    isDeleted: {
+        type: Boolean,
+    },
+}, { versionKey: false });
+
+schema.pre('save', async function save(next) {
+    if (!this.isModified('password')) {
+        return next();
+    }
+
+    try {
+        this.password = await bcrypt.hash(this.password, Number(process.env.JWT_SALT));
+    } catch (err) {
+        next(err);
+    }
+});
+
+export default schema;