+++ server/src/middleware/admin.middleware.js
@@ -1,12 +1,17 @@
-import isAdmin from '../helpers/isAdmin.js';
+import ConstMessages from 'asset-management-common/constants/constMessages.js';
+import ConstCodes from 'asset-management-common/constants/constCodes.js';
+import isAdmin from '../lib/helpers/isAdmin.js';
+import validateError from '../lib/helpers/validateError.js';
 
 export default async function (req, res, next) {
     try {
         if (!isAdmin(req.user)) {
-            throw new Error('Invalid permissions');
+            throw new Error(ConstMessages.invalidPermissions);
         }
         next();
     } catch (err) {
-        return res.status(403).json({ message: err.message });
+        res.status(ConstCodes.forbidden).send(
+            validateError(err) || ConstMessages.internalServerError,
+        );
     }
 }