+++ server/src/startup/dataGenerator/index.js
@@ -1,3 +1,3 @@
 import GenerateUsers from './generateUsers.js';
 
-export default async (environment) => [await GenerateUsers(environment)];
+export default async () => [await GenerateUsers()];